feat: enhance accounts overview with sorting and grouping features
- Introduced sorting functionality for accounts, allowing users to dynamically set sorting rules.
- Added a new sorting menu with drag-and-drop capabilities for better user experience.
- Updated the accounts overview to display sorted groups and ungrouped accounts effectively.
- Refactored related components to integrate the new sorting logic and improve overall performance.
- Adjusted localization files to reflect changes in UI labels for sorting and account views.
