+++ app/[locale]/dashboard/components/accounts/accounts-overview.tsx
@@ -8,9 +8,37 @@ import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogDescription, DialogFooter, DialogTrigger } from "@/components/ui/dialog"
 import { Sheet, SheetContent, SheetHeader, SheetTitle, SheetDescription, SheetFooter } from "@/components/ui/sheet"
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover"
+import {
+  Select,
+  SelectContent,
+  SelectItem,
+  SelectTrigger,
+  SelectValue,
+} from "@/components/ui/select"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
-import { CalendarIcon, Info, Plus, X, Clock, CheckCircle, XCircle, DollarSign, Trash2, Save, Settings, ChevronLeft, ChevronRight, Loader2, GripVertical, LayoutGrid, Table } from "lucide-react"
+import {
+  ArrowDown,
+  ArrowUp,
+  CalendarIcon,
+  Info,
+  ListOrdered,
+  Plus,
+  X,
+  Clock,
+  CheckCircle,
+  XCircle,
+  DollarSign,
+  Trash2,
+  Save,
+  Settings,
+  ChevronLeft,
+  ChevronRight,
+  Loader2,
+  GripVertical,
+  LayoutGrid,
+  Table,
+} from "lucide-react"
 import { Calendar } from "@/components/ui/calendar"
 import { format, Locale } from "date-fns"
 import { cn, calculateTradingDays } from "@/lib/utils"
@@ -39,6 +67,7 @@ import { useAccountOrderStore } from '@/store/account-order-store'
 import { useAccountsViewPreferenceStore } from '@/store/accounts-view-preference-store'
 import { savePayoutAction, removeAccountsFromTradesAction } from '@/server/accounts'
 import { useModalStateStore } from '@/store/modal-state-store'
+import { SortingState } from "@tanstack/react-table"
 import {
   DndContext,
   closestCenter,
@@ -97,6 +126,150 @@ interface PayoutDialogProps {
   onDelete?: () => Promise<void>
   isLoading?: boolean
   isDeleting?: boolean
+}
+
+type SortOption = {
+  id: string
+  label: string
+}
+
+function toValidDate(value: Date | string | null | undefined) {
+  if (!value) return null
+  const date = value instanceof Date ? value : new Date(value)
+  return Number.isNaN(date.getTime()) ? null : date
+}
+
+function getAccountStartDate(account: Account) {
+  const tradeDates = (account.trades ?? [])
+    .map((trade) => toValidDate(trade.entryDate))
+    .filter((date): date is Date => Boolean(date))
+    .sort((a, b) => a.getTime() - b.getTime())
+
+  if (tradeDates.length > 0) return tradeDates[0]
+
+  const dailyDates = (account.dailyMetrics ?? [])
+    .map((metric) => toValidDate(metric.date))
+    .filter((date): date is Date => Boolean(date))
+    .sort((a, b) => a.getTime() - b.getTime())
+
+  return dailyDates[0] ?? null
+}
+
+function getAccountBalance(account: Account) {
+  return account.metrics?.currentBalance ?? account.startingBalance ?? 0
+}
+
+function getAccountSortValue(account: Account, ruleId: string) {
+  switch (ruleId) {
+    case "account":
+      return account.number || ""
+    case "propfirm":
+      return account.propfirm || ""
+    case "startDate":
+      return getAccountStartDate(account)?.getTime() ?? Number.POSITIVE_INFINITY
+    case "funded":
+      return account.evaluation === false ? 1 : 0
+    case "balance":
+      return getAccountBalance(account)
+    case "targetProgress":
+      return account.metrics?.progress ?? 0
+    case "drawdown":
+      return account.metrics?.remainingLoss ?? 0
+    case "consistency":
+      if (!account.metrics?.hasProfitableData) return 0
+      return account.metrics.isConsistent || account.consistencyPercentage === 100
+        ? 2
+        : 1
+    case "maxDailyProfit":
+      return account.metrics?.highestProfitDay ?? 0
+    case "tradingDays":
+      return account.metrics?.totalTradingDays ?? 0
+    default:
+      return ""
+  }
+}
+
+function compareSortValues(a: unknown, b: unknown, desc: boolean) {
+  let result = 0
+  if (typeof a === "number" && typeof b === "number") {
+    result = a - b
+  } else {
+    result = String(a).localeCompare(String(b), undefined, {
+      sensitivity: "base",
+    })
+  }
+  return desc ? -result : result
+}
+
+function SortRuleItem({
+  sort,
+  label,
+  reorderLabel,
+  toggleLabel,
+  removeLabel,
+  onToggleDirection,
+  onRemove,
+}: {
+  sort: SortingState[number]
+  label: string
+  reorderLabel: string
+  toggleLabel: string
+  removeLabel: string
+  onToggleDirection: () => void
+  onRemove: () => void
+}) {
+  const { attributes, listeners, setNodeRef, transform, transition, isDragging } =
+    useSortable({ id: sort.id })
+  const style = {
+    transform: CSS.Transform.toString(transform),
+    transition,
+  }
+
+  return (
+    <div
+      ref={setNodeRef}
+      style={style}
+      className={cn(
+        "flex items-center gap-2 rounded-md border bg-background px-2 py-1.5 text-sm",
+        isDragging && "opacity-70 shadow-sm"
+      )}
+    >
+      <button
+        type="button"
+        className="cursor-grab text-muted-foreground active:cursor-grabbing"
+        {...attributes}
+        {...listeners}
+        aria-label={reorderLabel}
+      >
+        <GripVertical className="h-4 w-4" />
+      </button>
+      <span className="flex-1 truncate">{label}</span>
+      <Button
+        type="button"
+        variant="ghost"
+        size="icon"
+        onClick={onToggleDirection}
+        className="h-7 w-7"
+        aria-label={toggleLabel}
+      >
+        {sort.desc ? (
+          <ArrowDown className="h-4 w-4" />
+        ) : (
+          <ArrowUp className="h-4 w-4" />
+        )}
+      </Button>
+      <Button
+        type="button"
+        variant="ghost"
+        size="icon"
+        onClick={onRemove}
+        className="h-7 w-7 text-muted-foreground hover:text-foreground"
+        aria-label={removeLabel}
+      >
+        <X className="h-4 w-4" />
+      </Button>
+    </div>
+  )
 }
 
 
@@ -591,7 +764,32 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
   const [pendingChanges, setPendingChanges] = useState<Partial<Account> | null>(null)
   const [isSavingPayout, setIsSavingPayout] = useState(false)
   const [isDeletingPayout, setIsDeletingPayout] = useState(false)
+  const [sorting, setSorting] = useState<SortingState>([])
+  const [sortingMenuOpen, setSortingMenuOpen] = useState(false)
+  const [pendingSortId, setPendingSortId] = useState("")
   const shouldUpdateSelectedAccount = useRef(false)
+
+  const sortOptions = useMemo<SortOption[]>(
+    () => [
+      { id: "group", label: t("accounts.table.group") },
+      { id: "account", label: t("accounts.table.account") },
+      { id: "propfirm", label: t("accounts.table.propfirm") },
+      { id: "startDate", label: t("accounts.table.startDate") },
+      { id: "funded", label: t("accounts.table.funded") },
+      { id: "balance", label: t("accounts.table.balance") },
+      { id: "targetProgress", label: t("accounts.table.targetProgress") },
+      { id: "drawdown", label: t("accounts.table.drawdownRemaining") },
+      { id: "consistency", label: t("propFirm.card.consistency") },
+      { id: "maxDailyProfit", label: t("propFirm.card.highestDailyProfit") },
+      { id: "tradingDays", label: t("propFirm.card.tradingDays") },
+    ],
+    [t]
+  )
+
+  const availableSortOptions = useMemo(
+    () => sortOptions.filter((option) => !sorting.some((rule) => rule.id === option.id)),
+    [sortOptions, sorting]
+  )
 
   // Drag and drop sensors
   const sensors = useSensors(
@@ -694,6 +892,75 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
 
     return { filteredAccounts: configuredAccounts, unconfiguredAccounts }
   }, [trades, accounts, accountNumbers, groups])
+
+  const { sortedGroupEntries, sortedUngroupedAccounts } = useMemo(() => {
+    const groupSortRule = sorting.find((rule) => rule.id === "group")
+    const accountSortRules = sorting.filter((rule) => rule.id !== "group")
+
+    const sortAccounts = (groupId: string, groupAccounts: Account[]) => {
+      if (accountSortRules.length === 0) {
+        return getOrderedAccounts(groupId, groupAccounts) as Account[]
+      }
+      return [...groupAccounts].sort((a, b) => {
+        for (const rule of accountSortRules) {
+          const aValue = getAccountSortValue(a, rule.id)
+          const bValue = getAccountSortValue(b, rule.id)
+          const compare = compareSortValues(aValue, bValue, rule.desc)
+          if (compare !== 0) return compare
+        }
+        return 0
+      })
+    }
+
+    const groupEntries = groups
+      .map((group) => {
+        const groupAccounts = filteredAccounts.filter((account) =>
+          group.accounts.some((a) => a.number === account.number)
+        )
+        if (groupAccounts.length === 0) return null
+        return {
+          group,
+          accounts: sortAccounts(group.id, groupAccounts),
+        }
+      })
+      .filter((value): value is { group: typeof groups[number]; accounts: Account[] } =>
+        Boolean(value)
+      )
+
+    if (groupSortRule) {
+      groupEntries.sort((a, b) =>
+        compareSortValues(a.group.name, b.group.name, groupSortRule.desc)
+      )
+    } else if (accountSortRules.length > 0) {
+      groupEntries.sort((a, b) => {
+        const aAccount = a.accounts[0]
+        const bAccount = b.accounts[0]
+        if (!aAccount && !bAccount) return 0
+        if (!aAccount) return 1
+        if (!bAccount) return -1
+        for (const rule of accountSortRules) {
+          const aValue = getAccountSortValue(aAccount, rule.id)
+          const bValue = getAccountSortValue(bAccount, rule.id)
+          const compare = compareSortValues(aValue, bValue, rule.desc)
+          if (compare !== 0) return compare
+        }
+        return 0
+      })
+    }
+
+    const groupedAccountNumbers = new Set(
+      groups.flatMap((group) => group.accounts.map((a) => a.number))
+    )
+    const ungroupedAccounts = filteredAccounts.filter(
+      (account) => !groupedAccountNumbers.has(account.number ?? "")
+    )
+    const sortedUngrouped = sortAccounts("ungrouped", ungroupedAccounts)
+
+    return {
+      sortedGroupEntries: groupEntries,
+      sortedUngroupedAccounts: sortedUngrouped,
+    }
+  }, [filteredAccounts, getOrderedAccounts, groups, sorting])
 
   const dailyMetrics = useMemo(() => {
     if (!selectedAccountForTable) return []
@@ -903,34 +1170,169 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
                 {t("filters.manageAccounts")}
               </span>
             </Button>
-            <Button
-              variant={view === "cards" ? "default" : "outline"}
-              size="sm"
-              onClick={() => setView("cards")}
-              className={cn(
-                "gap-1.5",
-                size === "small" ? "h-7 px-2 text-xs" : "h-8"
-              )}
-            >
-              <LayoutGrid className="h-3.5 w-3.5" />
-              <span className={cn(size === "small" && "sr-only")}>
-                {t("accounts.view.cards")}
-              </span>
-            </Button>
-            <Button
-              variant={view === "table" ? "default" : "outline"}
-              size="sm"
-              onClick={() => setView("table")}
-              className={cn(
-                "gap-1.5",
-                size === "small" ? "h-7 px-2 text-xs" : "h-8"
-              )}
-            >
-              <Table className="h-3.5 w-3.5" />
-              <span className={cn(size === "small" && "sr-only")}>
-                {t("accounts.view.table")}
-              </span>
-            </Button>
+            <Popover open={sortingMenuOpen} onOpenChange={setSortingMenuOpen}>
+              <PopoverTrigger asChild>
+                <Button
+                  variant="outline"
+                  size="sm"
+                  className={cn(
+                    "gap-1.5",
+                    size === "small" ? "h-7 px-2 text-xs" : "h-8"
+                  )}
+                >
+                  <ListOrdered className="h-3.5 w-3.5" />
+                  <span className={cn(size === "small" && "sr-only")}>
+                    {sorting.length > 0
+                      ? t("table.sortingRules", { count: sorting.length })
+                      : t("table.sorting")}
+                  </span>
+                </Button>
+              </PopoverTrigger>
+              <PopoverContent align="end" className="w-80 p-3">
+                <div className="space-y-3">
+                  {sorting.length === 0 ? (
+                    <div className="text-sm text-muted-foreground">
+                      {t("table.noSorting")}
+                    </div>
+                  ) : (
+                    <DndContext
+                      sensors={sensors}
+                      collisionDetection={closestCenter}
+                      onDragEnd={(event: DragEndEvent) => {
+                        const { active, over } = event
+                        if (!over || active.id === over.id) return
+                        const oldIndex = sorting.findIndex(
+                          (rule) => rule.id === active.id
+                        )
+                        const newIndex = sorting.findIndex(
+                          (rule) => rule.id === over.id
+                        )
+                        if (oldIndex === -1 || newIndex === -1) return
+                        setSorting((prev) => arrayMove(prev, oldIndex, newIndex))
+                      }}
+                    >
+                      <SortableContext
+                        items={sorting.map((rule) => rule.id)}
+                        strategy={verticalListSortingStrategy}
+                      >
+                        <div className="space-y-2 max-h-48 overflow-y-auto pr-1">
+                          {sorting.map((rule) => {
+                            const label =
+                              sortOptions.find(
+                                (option) => option.id === rule.id
+                              )?.label ?? rule.id
+                            return (
+                              <SortRuleItem
+                                key={rule.id}
+                                sort={rule}
+                                label={label}
+                                reorderLabel={t("table.reorderSort")}
+                                toggleLabel={
+                                  rule.desc
+                                    ? t("table.sortDescending")
+                                    : t("table.sortAscending")
+                                }
+                                removeLabel={t("table.removeSort")}
+                                onToggleDirection={() =>
+                                  setSorting((prev) =>
+                                    prev.map((item) =>
+                                      item.id === rule.id
+                                        ? { ...item, desc: !item.desc }
+                                        : item
+                                    )
+                                  )
+                                }
+                                onRemove={() =>
+                                  setSorting((prev) =>
+                                    prev.filter((item) => item.id !== rule.id)
+                                  )
+                                }
+                              />
+                            )
+                          })}
+                        </div>
+                      </SortableContext>
+                    </DndContext>
+                  )}
+                  <div className="flex items-center gap-2">
+                    <Select
+                      value={pendingSortId}
+                      onValueChange={(value) => {
+                        const nextValue = value === "__none" ? "" : value
+                        setPendingSortId(nextValue)
+                        if (nextValue) {
+                          setSorting((prev) => [
+                            ...prev,
+                            { id: nextValue, desc: false },
+                          ])
+                          setPendingSortId("")
+                        }
+                      }}
+                    >
+                      <SelectTrigger className="h-8 flex-1">
+                        <SelectValue placeholder={t("table.pickSortColumn")} />
+                      </SelectTrigger>
+                      <SelectContent>
+                        {availableSortOptions.length === 0 ? (
+                          <SelectItem value="__none" disabled>
+                            {t("table.noMoreSortOptions")}
+                          </SelectItem>
+                        ) : (
+                          availableSortOptions.map((option) => (
+                            <SelectItem key={option.id} value={option.id}>
+                              {option.label}
+                            </SelectItem>
+                          ))
+                        )}
+                      </SelectContent>
+                    </Select>
+                  </div>
+                  <div className="flex items-center justify-between">
+                    <Button
+                      variant="ghost"
+                      size="sm"
+                      onClick={() => setSorting([])}
+                      disabled={sorting.length === 0}
+                    >
+                      {t("table.clearSorting")}
+                    </Button>
+                  </div>
+                </div>
+              </PopoverContent>
+            </Popover>
+            <Tabs value={view} onValueChange={(value) => setView(value as "cards" | "table")}>
+              <TabsList
+                className={cn(
+                  "gap-1",
+                  size === "small" ? "h-7 px-1" : "h-8 px-1"
+                )}
+              >
+                <TabsTrigger
+                  value="cards"
+                  className={cn(
+                    "gap-1.5",
+                    size === "small" ? "h-6 px-2 text-xs" : "h-7"
+                  )}
+                >
+                  <LayoutGrid className="h-3.5 w-3.5" />
+                  <span className={cn(size === "small" && "sr-only")}>
+                    {t("accounts.view.charts")}
+                  </span>
+                </TabsTrigger>
+                <TabsTrigger
+                  value="table"
+                  className={cn(
+                    "gap-1.5",
+                    size === "small" ? "h-6 px-2 text-xs" : "h-7"
+                  )}
+                >
+                  <Table className="h-3.5 w-3.5" />
+                  <span className={cn(size === "small" && "sr-only")}>
+                    {t("accounts.view.table")}
+                  </span>
+                </TabsTrigger>
+              </TabsList>
+            </Tabs>
           </div>
         </div>
       </CardHeader>
@@ -1000,20 +1402,7 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
           {view === "cards" ? (
             <div className="mt-4">
               <div className="space-y-6">
-                {groups.map((group, groupIndex) => {
-                  // Filter accounts for this group
-                  const groupAccounts = filteredAccounts.filter(account => {
-                    console.log('account', JSON.stringify(account, null, 2))
-                    // Find the account in the group's accounts
-                    return group.accounts.some(a => a.number === account.number);
-                  });
-
-                  // Skip groups with no accounts
-                  if (groupAccounts.length === 0) return null;
-
-                  // Get ordered accounts for this group
-                  const orderedAccounts = getOrderedAccounts(group.id, groupAccounts) as Account[];
-
+                {sortedGroupEntries.map(({ group, accounts: orderedAccounts }, groupIndex) => {
                   // Generate a consistent color for each group based on group index
                   const groupColors = [
                     'border-blue-200/50 bg-blue-50/30 dark:border-blue-800/30 dark:bg-blue-950/20',
@@ -1042,7 +1431,7 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
                             {group.name}
                           </h3>
                           <div className="text-xs text-muted-foreground bg-white/50 dark:bg-black/20 px-2 py-1 rounded-full">
-                            {groupAccounts.length} {groupAccounts.length === 1 ? 'account' : 'accounts'}
+                            {orderedAccounts.length} {orderedAccounts.length === 1 ? 'account' : 'accounts'}
                           </div>
                         </div>
                       </div>
@@ -1081,19 +1470,7 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
 
                 {/* Show ungrouped accounts */}
                 {(() => {
-
-                  const groupedAccountNumbers = new Set(
-                    groups.flatMap(group => group.accounts.map(a => a.number))
-                  );
-
-                  const ungroupedAccounts = filteredAccounts.filter(
-                    account => !groupedAccountNumbers.has(account.number ?? '')
-                  );
-
-                  if (ungroupedAccounts.length === 0) return null;
-
-                  // Get ordered accounts for ungrouped (using a special key)
-                  const orderedUngroupedAccounts = getOrderedAccounts('ungrouped', ungroupedAccounts) as Account[];
+                  if (sortedUngroupedAccounts.length === 0) return null;
 
                   return (
                     <div
@@ -1109,7 +1486,7 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
                             {t('propFirm.ungrouped')}
                           </h3>
                           <div className="text-xs text-muted-foreground bg-white/50 dark:bg-black/20 px-2 py-1 rounded-full">
-                            {ungroupedAccounts.length} {ungroupedAccounts.length === 1 ? 'account' : 'accounts'}
+                            {sortedUngroupedAccounts.length} {sortedUngroupedAccounts.length === 1 ? 'account' : 'accounts'}
                           </div>
                         </div>
                       </div>
@@ -1123,11 +1500,11 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
                           modifiers={[restrictToHorizontalAxis]}
                         >
                           <SortableContext
-                            items={orderedUngroupedAccounts.map(acc => acc.number)}
+                            items={sortedUngroupedAccounts.map(acc => acc.number)}
                             strategy={horizontalListSortingStrategy}
                           >
                             <div className="flex gap-3 overflow-x-auto pb-2 min-h-fit">
-                              {orderedUngroupedAccounts.map(account => {
+                              {sortedUngroupedAccounts.map(account => {
                                 if (!account.number) return null;
                                 return (
                                   <DraggableAccountCard
@@ -1152,6 +1529,8 @@ export function AccountsOverview({ size }: { size: WidgetSize }) {
               accounts={filteredAccounts}
               groups={groups}
               onSelectAccount={(account) => setSelectedAccountForTable(account)}
+              sorting={sorting}
+              onSortingChange={setSorting}
             />
           )}
         </div>
